+++ packages/emoji/src/components/common/ToneSelector.tsx
@@ -28,7 +28,7 @@ export default class ToneSelector extends PureComponent<Props, undefined> {
             <EmojiButton
               key={`${tone.id}`}
               // tslint:disable-next-line:jsx-no-lambda
-              onSelected={() => onToneSelected(i + 1)}
+              onSelected={() => onToneSelected(i)}
               emoji={tone}
             />
           ))