+++ packages/emoji/src/components/picker/EmojiPickerFooter.tsx
@@ -5,14 +5,15 @@ import * as classNames from 'classnames';
 import * as styles from './styles';
 import EmojiPreview from '../common/EmojiPreview';
 import EmojiUploadPicker, { OnUploadEmoji } from '../common/EmojiUploadPicker';
-import { EmojiDescription, OnToneSelected } from '../../types';
+import { EmojiDescription, EmojiDescriptionWithVariations, OnToneSelected, ToneSelection } from '../../types';
 
 export interface Props {
   selectedEmoji?: EmojiDescription;
-  selectedTone?: number;
+  selectedTone?: ToneSelection;
   onToneSelected?: OnToneSelected;
   uploading: boolean;
   initialUploadName?: string;
+  toneEmoji?: EmojiDescriptionWithVariations;
   uploadErrorMessage?: string;
   onUploadCancelled: () => void;
   onUploadEmoji: OnUploadEmoji;
@@ -27,6 +28,7 @@ export default class EmojiPickerFooter extends PureComponent<Props, undefined> {
       onUploadEmoji,
       selectedEmoji,
       selectedTone,
+      toneEmoji,
       uploadErrorMessage,
       uploading,
     } = this.props;
@@ -53,6 +55,7 @@ export default class EmojiPickerFooter extends PureComponent<Props, undefined> {
       <div className={previewFooterClassnames}>
         <EmojiPreview
           emoji={selectedEmoji}
+          toneEmoji={toneEmoji}
           selectedTone={selectedTone}
           onToneSelected={onToneSelected}
         />